project-one/requirements.txt
pymongo[srv,tls]
python-dotenv
requests
sentence-transformers
